Transaction History and export CSV function completed The issue in calling user_id also fixed
