+++ src/topic_2_ledger_system/exported/TransactionHistory_11/2024.csv
@@ -1,2 +1,2 @@
-"%-20s %-15s %-10s %-10s %-20s %-15s\n", "Transaction ID", "User ID", "Type", "Amount", "Description", "Date"
-3                    2               debit      2000       Salary               11/11/2024          
+Transaction ID", "User ID", "Type", "Amount", "Description", "Date"
+3, 2, debit, 2000, Salary, 11/11/2024
