+++ src/topic_2_ledger_system/main_system.java
@@ -23,6 +23,19 @@ public class main_system {
      */
     public static void main(String[] args) throws IOException{
         login_register.initialize();
+        
+        //Call the method for Transaction History
+        try {
+            // Initialize login or registration process
+            int user_id = login_register.initialize();
 
+            // Get the month and year from the user
+            Scanner scanner = new Scanner(System.in);
+            System.out.print("Enter the month and year (MM-YYYY): ");
+            String monthYear = scanner.nextLine();
+            view_export_csv.viewAndExportTransactions(user_id, monthYear);
+        } catch (IOException e) {
+            System.err.println("An error occurred: " + e.getMessage());
+        }
     }
 }
